excel_work/requirements.txt
asgiref==3.8.1
Django==5.0.3
sqlparse==0.4.4
tzdata==2024.1

djangorestframework~=3.15.1